aaa_base go-away patches, faster libmpkg shutdown
git-svn-id: svn://agilialinux.ru/mpkg/mpkg/trunk@198 32a6d772-3a6b-4160-9f27-08b1a0ab14b6

+++ mpkgsupport/file_routines.cpp
@@ -17,7 +17,9 @@ TempFileController::TempFileController() {
 }
 
 TempFileController::~TempFileController() {
+	printf("Cleaning %d temporary files, please wait\n", tFiles.size());
 	clear_all();
+	printf("Done\n");
 }
 
 int getProcessPid(const string& name) {
@@ -47,7 +49,7 @@ void TempFileController::clear_all() {
 	for (size_t i=0; i<tFiles.size(); ++i) {
 		if (access(tFiles[i].c_str(), F_OK)) unlink(tFiles[i].c_str()); // If plain file
 		else if (access(string(tFiles[i]+".gz").c_str(), F_OK)) unlink(string(tFiles[i]+".gz").c_str()); // Sometimes we use gz extension
-		system("rm -rf " + tFiles[i]); // And sometimes it is a directory
+		if (FileExists(tFiles[i])) system("rm -rf " + tFiles[i]); // And sometimes it is a directory
 	}
 	tFiles.clear(); // Clean-up list - for future use
 }
